requirements.txt
itemadapter==0.8.0
mysql_connector_repackaged==0.3.1
Scrapy==2.9.0
